Move the cloud config helpers into the enterprise packages
