+++ pkg/controller/intrusiondetection/intrusiondetection_controller.go
@@ -434,7 +434,7 @@ func (r *ReconcileIntrusionDetection) Reconcile(ctx context.Context, request rec
 	if r.opts.Cloud && r.opts.ElasticExternal && !r.opts.MultiTenant {
 		// For Calico Cloud single-tenant clusters sharing an external ES, extract the tenant
 		// information from the cloud config map.
-		cloudConfig, err := utils.GetCloudConfig(ctx, r.client)
+		cloudConfig, err := eutils.GetCloudConfig(ctx, r.client)
 		if err != nil {
 			r.status.SetDegraded(operatorv1.ResourceReadError, "Failed to read cloud config", err, reqLogger)
 			return reconcile.Result{}, err
