+++ pkg/enterprise/cloudconfig/cloudconfig_test.go
